Feedback from Ashwin: Name change from
Nedunoori, Ashwin to
------------------------------->>>>>
Nedunoori, Ashwin Krishna

.............................................


TRUST changes:

page 25

2.3
Trust "is"
....................................
"Without trust, an organization is not meet its quality standard
standard."------------------>>>>

"Without trust, an organization will not meet its quality standart"

-------------------------------------------


page 26

In the figure 2.6,  an ARROW (------>) is missing from SECURITY AGENT II to NODE
B

2.3.3
Page 28
Trust value equation :

Initial trust should come before  our recommended equation (2.3.4)

Initaial trust equation:
The artifact of this decision is in the node notation of Equation (2.7)

page 29

The initial intent was once the trust value is calculated, the trust value of
the nodes can be
increased or reduced according to Equation (2.8)

page 30

The value1 can be calculated from (2.2) having the list of documents and the
summation
of the values ---> list    value of the particular node.
-----------------------------------
should be changed to
------------------------------------
The value1 of the (TABLE 2.2) can be calculated from (TABLE 2.3) having the list
of documents and the summation
of the values --->list   value of the particular node.

2.3.5
page 30

However, to trust must also have a method to increase otherwise, over time, all
nodes will be deemed untrustworthy.
---------------------------------->>>>>

However, Trust should also have a method to increase its trust value otherwise,
overtime, all nodes will be deemed untrustworthy.

--------------------------------------------------------

page 31
2.3.6  Trust value calculation

The (Figure 3 ) should be changed to

The (Figure 2.8) below